add outer clothing slot and first armor
