+++ Resources/Prototypes/_CE/Entities/Clothing/Head/helmets.yml
@@ -58,7 +58,8 @@
   - type: Clothing
     sprite: _CE/Clothing/Head/Helmets/brass_insulator.rsi
   - type: CEEnergyRadiationArmor
-    armor: 0.3
+    armor: 0.25
+  - type: MultiHandedItem
   - type: HideLayerClothing
     slots:
     - Hair
@@ -67,7 +68,6 @@
   - type: Armor
     modifiers:
       coefficients:
-        Blunt: 0.9
-        Slash: 0.9
-        Piercing: 0.9
-        Shock: 0.75
+        Shock: 0.9
+  - type: StaticPrice
+    price: 100
