+++ Resources/Prototypes/_CE/Entities/Clothing/OuterClothing/base.yml
@@ -1,0 +1,16 @@
+- type: entity
+  abstract: true
+  parent: Clothing
+  id: CEClothingOuterClothingBase
+  categories: [ ForkFiltered ]
+  description: Armor or something important to wear.
+  components:
+  - type: Clothing
+    equipDelay: 2.5
+    unequipDelay: 1.5
+    slots:
+    - outerClothing
+  - type: Sprite
+    state: icon
+  - type: Item
+    size: Huge
