+++ Resources/Prototypes/_CE/InventoryTemplates/carcat_inventory_template.yml
@@ -56,6 +56,13 @@
 #      blacklist:
 #        tags:
 #          - PetOnly
+    - name: outerClothing
+      slotTexture: suit
+      slotFlags: OUTERCLOTHING
+      stripTime: 6
+      uiWindowPos: 2,2
+      strippingWindowPos: 1,2
+      displayName: Suit
     - name: cloak
       slotTexture: cloak
       slotFlags: CLOAK
